refactoring

+++ src/disasm.c
@@ -18,7 +18,7 @@
 //
 // *******************************************************************************************
 
-int disasm(uint16_t pc, uint8_t *RAM, char *line, unsigned int max_line, uint8_t bank, uint8_t implied_status, int32_t *eff_addr) {
+int disasm(uint16_t pc, uint8_t *RAM, char *line, unsigned int max_line, int16_t bank, uint8_t implied_status, int32_t *eff_addr) {
 	uint8_t opcode = debug_read6502(pc, bank);
 	const char *mnemonic = regs.is65c816 ? mnemonics_c816[opcode] : mnemonics_c02[opcode];
